pyproject.toml
[project]
name = "project-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "black>=25.9.0",
    "fastapi>=0.120.0",
    "httpx>=0.28.1",
    "polyfactory>=3.1.0",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.3",
    "pyjwt>=2.10.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-multipart>=0.0.20",
    "ruff>=0.14.2",
    "sqlalchemy>=2.0.44",
    "time-machine>=3.1.0",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
pythonpath = "src"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term",
    "--cov-report=html",
]


[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
